+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Fixed reach not applicable to chestplates
+* Fixed blood pact missing lang
 * Fixed obsidian repair kit being too dark
 
 ## 1.11.0